[ER-109] - Incrementer now operates with Long.

+++ test/src/main/java/neuro/test/OffsetIncrementer.kt
@@ -4,7 +4,7 @@ class OffsetIncrementer(
   private val incrementer: Incrementer = Incrementer(),
   private val interval: Int = 1000
 ) {
-  fun getAndIncrement(): Int {
+  fun getAndIncrement(): Long {
     return incrementer.getAndIncrement() * interval
   }
 }
